+++ commands/textureTools/instapass.js
@@ -5,129 +5,119 @@ const fs      = require('fs');
 const fetch   = require('node-fetch');
 const strings = require('../../res/strings');
 const colors  = require('../../res/colors');
+const settings = require('../../settings.js');
 
-const REPOSITORIES = [
-	'Compliance-Java-32x',
-	'Compliance-Java-64x',
-	'Compliance-Bedrock-32x',
-	'Compliance-Bedrock-64x'
-]
-
-const BE_BRANCHES = [ '1.16.210' ];
-const JE_BRANCHES = [ '1.12.2', '1.13.2', '1.14.4', '1.15.2', '1.16.5', '1.17' ];
-
-const { parseArgs } = require('../../functions/utility/parseArgs.js');
 const { date }      = require('../../functions/utility/date.js');
 const { warnUser }  = require('../../functions/warnUser.js');
 const { doPush }    = require('../../functions/doPush.js');
 const { jsonContributionsJava, jsonContributionsBedrock } = require('../../helpers/fileHandler');
 
+const BE_BRANCHES = [ '1.16.210' ];
+const JE_BRANCHES = [ '1.12.2', '1.13.2', '1.14.4', '1.15.2', '1.16.5', '1.17' ];
+
 module.exports = {
-	name: 'push',
-	description: strings.HELP_DESC_PUSH,
+	name: 'instapass',
+	description: strings.HELP_DESC_INSTAPASS,
 	guildOnly: false,
 	uses: strings.COMMAND_USES_MODS,
-	syntax: `${prefix}push -r -p -a + file attached`,
-	flags: '-r | --repo :\n\tCompliance-[Java|Bedrock]-[32x-64x]\n-p | --path :\n\tTexture path \n-a | --author :\n\tDiscord tag of texture\'s author.',
-	example: `${prefix}push -r=Compliance-Java-32x -p=textures/block/stone.png -a=Someone#1234`,
+	syntax: `${prefix}instapass <message URL>`,
+	flags: 'No flags',
+	example: `${prefix}instapass https://discord.com/channels/.../.../...`,
 	async execute(client, message, args) {
 
 		if(!message.member.hasPermission('ADMINISTRATOR')) return warnUser(message,strings.COMMAND_NO_PERMISSION);
-		
-		args = parseArgs(args);
-		//console.log(args)
 
-		var haveAuth = false;
-		var havePath = false;
-		var haveRepo = false;
+		let valURL = args[0];
 
-		let valAuth;
-		let valPath;
-		let valRepo;
-		let valIndex;
-		let valType;
-		let valSize;
+		if (valURL.startsWith('https://discord.com/channels')) {
+			message.channel.messages.fetch(valURL.split('/').pop()).then(async msg => {
+				if (msg.attachments.size > 0) {
 
-		let fileHandle;
-		let textures;
+					let folder;
+					let name;
+					let textures;
+					let valSize;
+					let valType;
+					let valRepo;
+					let valIndex;
 
-		// Check args:
-		for (var i in args) {
-			if (args[i].startsWith('-a=') || args[i].startsWith('--author=')) {
-				haveAuth = true; 
-				valAuth = args[i].replace('-a=', '').replace('--author=', '');
-			}
-			if (args[i].startsWith('-p=') || args[i].startsWith('--path=')) {
-				havePath = true; 
-				valPath = args[i].replace('-p=', '').replace('--path=', '');
-			}
-			if (args[i].startsWith('-r=') || args[i].startsWith('--repo=')) {
-				haveRepo = true;
-				valRepo = args[i].replace('-r=', '').replace('--repo=');
-			}
+					args = msg.content.split(/ +/);
+					
+					folder = args[1].split('[').pop().split(']')[0];
+					name   = args[0];
+
+					if (msg.channel.id === settings.C32_SUBMIT_1 || msg.channel.id === settings.C64_SUBMIT_1) valType = 'java';
+					else if (msg.channel.id === settings.C32_SUBMIT_1B || msg.channel.id === settings.C64_SUBMIT_1B) valType = 'bedrock';
+
+					if (name === undefined || name == '') name = 'Not Provided';
+					if (!valType) valType = 'java';
+
+					if (message.guild.id === settings.C32_ID) {
+						valSize = 32;
+						if (valType == 'java') valRepo = 'Compliance-Java-32x';
+						if (valType == 'bedrock') valRepo = 'Compliance-Bedrock-32x';
+					}
+					if (message.guild.id === settings.C64_ID) {
+						valSize = 64;
+						if (valType == 'java') valRepo = 'Compliance-Java-64x';
+						if (valType == 'bedrock') valRepo = 'Compliance-Bedrock-64x';
+					}
+
+					// debug:
+					valRepo = 'Compliance-Java-32x';
+					valSize = 32;
+
+					let valPath = `${folder}/${name}.png`;
+
+					if (valType == 'java') {
+						textures = await jsonContributionsJava.read();
+
+						for (const i in textures) {
+							if (textures[i].version[strings.LATEST_MC_JE_VERSION].includes(valPath)) {
+								valIndex = i;
+								break;
+							}
+						}
+
+						jsonContributionsJava.release();
+					}
+					else if (valType == 'bedrock') {
+						textures = await jsonContributionsBedrock.read();
+
+						for (const i in textures) {
+							if (textures[i].version[strings.LATEST_MC_BE_VERSION].includes(valPath)) {
+								valIndex = i;
+								break;
+							}
+						}
+
+						jsonContributionsBedrock.release();
+					}
+
+					if (!valIndex) return warnUser(message, `Can't find ${valPath} inside JSON.`);
+
+					await setAuthor(valType, valIndex, `${msg.author.id}`, valSize);
+					await download(msg, valType, valIndex, valRepo, valSize);
+					await doPush(`Instapass for ${valPath.split('/').pop()} executed by: ${message.author.username}`);
+					await msg.react('⏩');
+
+					const embed = new Discord.MessageEmbed()
+						.setTitle(`Instapassed texture:`)
+						.setColor(colors.COUNCIL)
+						.setDescription(`The council have decided to instapassed this [texture](${valURL})`);
+
+					await message.channel.send(embed);
+					if (!message.deleted) message.delete();
+
+				} else return warnUser(message, 'This message does not have any texture attached.');
+			}).catch(error => {
+				console.log('\n\n ------------------ INSTAPASS ERROR ------------------\n');
+				console.error(error);
+				console.log('\n -----------------------------------------------------');
+				return warnUser(message, `${error}.\nI can only fetch images from the same channel. Don\'t ask why, I don\'t know myself.`)
+			});
 		}
 
-		var warnMessage = '';
-		if (!haveAuth) warnMessage += 'No author given! ';
-		if (!havePath) warnMessage += 'No texture path given! ';
-		if (!haveRepo) warnMessage += 'No repository given! ';
-		if (message.attachments.size == 0) warnMessage += 'You did not attached the texture! ';
-		if (warnMessage != '') return warnUser(message, warnMessage);
-
-		// Check repository:
-		if (!REPOSITORIES.includes(valRepo)) return warnUser(message, 'This repository does not exist or is not supported');
-		if (REPOSITORIES.indexOf(valRepo) == 0 || REPOSITORIES.indexOf(valRepo) == 2) valSize = 32;
-		else valSize = 64;
-
-		// fetch author ID:
-		try {
-			valAuth = client.users.cache.find(u => u.tag === valAuth).id;
-		} catch(error) {
-			console.log('\n\n -------------- USER NOT FOUND IN CACHE --------------\n');
-			console.error(error);
-			console.log('\n -----------------------------------------------------');
-			return warnUser(message, 'User not found in cache');
-		}
-
-		// Start search trough JAVA textures:
-		fileHandle = jsonContributionsJava;
-		textures   = await fileHandle.read();
-
-		for (const i in textures) {
-			if (textures[i].version[strings.LATEST_MC_JE_VERSION].includes(valPath)) {
-				valIndex = i;
-				valType  = 'java';
-				break;
-			}
-		}
-
-		fileHandle.release();
-
-		// Start search trough BEDROCK textures:
-		if (!valType) {
-			fileHandle = jsonContributionsBedrock;
-			textures   = await fileHandle.read();
-
-			for (const i in textures) {
-				if (textures[i].version[strings.LATEST_MC_BE_VERSION].includes(valPath)) {
-					valIndex = i;
-					valType  = 'bedrock';
-					break;
-				}
-			}
-
-			fileHandle.release();
-		}
-		
-		if (!valType) return warnUser(message, 'Can\'t find this textures!')
-		else {
-			await setAuthor(valType, valIndex, valAuth, valSize);
-			await download(message, valType, valIndex, valRepo, valSize);
-			await doPush(`Manual Push for ${valPath.split('/').pop()} executed by: ${message.author.username}`);
-			await message.react('✅');
-		}
-
-		// console.log(valAuth, valRepo, valPath, valIndex, valType, valSize);
 	}
 }