+++ functions/utility/date.js
@@ -3,7 +3,7 @@ function date() {
 	var dd = String(today.getDate()).padStart(2, '0');
 	var mm = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
 	var yyyy = today.getFullYear();
-	return mm + '/' + dd + '/' + yyyy;
+	return dd + '/' + mm + '/' + yyyy;
 }
 
 exports.date = date;